% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSDindecesFunctions.R
\name{auc_index}
\alias{auc_index}
\title{PCSD AUC Index -- auc_index()}
\usage{
auc_index(scn)
}
\arguments{
\item{scn}{}
}
\description{
This function calculates the area under the PCSD curve for each
construct.
}
