% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WimpIndecesFunctions.R
\name{betw_index}
\alias{betw_index}
\title{Betweeness index -- betw_index()}
\usage{
betw_index(wimp, norm = TRUE)
}
\arguments{
\item{wimp}{Subject's Weigthed ImpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{norm}{If TRUE, the values will be standardized. Default is TRUE.}
}
\value{
Returns a vector with the betweeness index for each of the
constructs.
}
\description{
Function that calculates the betweenness of each of the
constructs. This is the number of times a geodesic path (shortest path)
between two other constructs passes through that construct in the digraph.
}
