% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WimpIndicesFunctions.R
\name{close_index}
\alias{close_index}
\title{Closeness index -- close_index()}
\usage{
close_index(wimp, norm = TRUE)
}
\arguments{
\item{wimp}{Subject's Weigthed ImpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{norm}{If TRUE, the values will be standardized. Default is TRUE.}
}
\value{
Returns a vector with the closeness index for each of the
constructs.
}
\description{
Function to calculate the closeness of a construct to the rest
of the constructs within the digraph.
}
