% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WimpIndecesFunctions.R
\name{dismatrix}
\alias{dismatrix}
\title{Distance Matrix -- dismatrix()}
\usage{
dismatrix(wimp, mode = "out")
}
\arguments{
\item{wimp}{Subject's Weigthed ImpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}

\item{mode}{Method to calculate the distances depending on the direction of
the edges.With "out" we calculate them respecting the direction of the edges,
"in" through the inverse of the direction of the edges and "all" without
taking into account the direction.}
}
\value{
Returns the digraph distance matrix. Matrix that contains the
distances of the shortest paths from one construct to another.
}
\description{
Function that calculates the shortest distance between each of
the pairs of digraph constructions.
}
