% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{idealdigraph}
\alias{idealdigraph}
\title{Ideal digraph}
\usage{
idealdigraph(
  wimp,
  inc = FALSE,
  layout = "circle",
  vertex.size = 1,
  edge.width = 1,
  color = "red/green"
)
}
\arguments{
\item{color}{}
}
\description{
Ideal digraph
}
