% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WimpIndecesFunctions.R
\name{inc_index}
\alias{inc_index}
\title{Ideal Inconsitencies -- inc_index()}
\usage{
inc_index(wimp)
}
\arguments{
\item{wimp}{Subject's Weigthed ImpGrid. It must be a "wimp" S3 object
imported by the \code{\link{importwimp}} function.}
}
\value{
WIP.
}
\description{
WIP
}
