% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{pcsd}
\alias{pcsd}
\title{Personal Constructs System Dynamics plot -- pcsd()}
\usage{
pcsd(scn, vline = NA)
}
\arguments{
\item{scn}{}

\item{vline}{}
}
\value{
Interactive plot created with plotly.
}
\description{
Interactive line plot of personal constructs system dinamics.
Show \code{\link{scenariomatrix}} values expressed in terms of distance to
Ideal-Self for each personal construct across the mathematical iterations.
}
