% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{pcsd_derivative}
\alias{pcsd_derivative}
\title{PCSD derivative -- pcsd_derivative()}
\usage{
pcsd_derivative(scn)
}
\arguments{
\item{scn}{}
}
\value{
Return a plot create via plotly r-package.
}
\description{
This function represents the first derivative for each of the
PCSD curves.
}
