% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSDindecesFunctions.R
\name{pcsd_summary}
\alias{pcsd_summary}
\title{PCSD summary -- pcsd_summary()}
\usage{
pcsd_summary(scn)
}
\arguments{
\item{scn}{}
}
\value{
Returns a matrix with the PCSD summary.
}
\description{
This function returns a summary of the PCSD. It informs us the
initial and final value of each construct and the difference between them.
}
