% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationFunctions.R
\name{scenariomatrix}
\alias{scenariomatrix}
\title{Scenario Matrix}
\usage{
scenariomatrix(
  wimp,
  act.vector,
  infer = "linear transform",
  thr = "linear",
  max.iter = 30,
  e = 1e-04,
  stop.iter = 3
)
}
\arguments{
\item{quiet}{}
}
\description{
Scenario Matrix
}
