% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{selfdigraph}
\alias{selfdigraph}
\title{Self digraph}
\usage{
selfdigraph(
  wimp,
  layout = "circle",
  vertex.size = 1,
  edge.width = 1,
  color = "red/green"
)
}
\arguments{
\item{color}{}
}
\description{
Self digraph
}
