% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GraphFunctions.R
\name{simdigraph}
\alias{simdigraph}
\title{Title}
\usage{
simdigraph(
  scn,
  niter = 0,
  layout = "graphopt",
  vertex.size = 1,
  edge.width = 1,
  color = "red/green"
)
}
\arguments{
\item{color}{}
}
\description{
Title
}
