% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCSDindicesFunctions.R
\name{stability_index}
\alias{stability_index}
\title{PCSD Stability Index -- stability_index()}
\usage{
stability_index(scn)
}
\arguments{
\item{scn}{}
}
\value{
Returns a vector with the standard deviation of each of the
constructs.
}
\description{
This function returns the standard deviation for each
construct over the mathematical iterations of the PCSD.
}
